feature: enable user to get signed url from aws (fixed failing specs)
